requirements.txt
celery==5.5.2
Django==5.2.1
djangorestframework==3.16.0

requests~=2.32.3
cryptography~=45.0.4